+++ app/(route)/(admin)/[id]/(_components)/Alim.tsx
