+++ app/(route)/(admin)/[id]/(_components)/AlimTable.tsx
@@ -4,7 +4,7 @@ import { flexRender } from "@tanstack/react-table";
 
 import { useAlimTableContext } from "../(hooks)/useAlimTable";
 import { AlimTHeaderColumn } from "../(constants)/AlimColumn";
-import { Pagination } from "../../../ui";
+import { Pagination } from "../../../../ui";
 
 export function AlimTable() {
   const { alimTable } = useAlimTableContext();
