+++ app/(route)/(admin)/[id]/(_components)/TeacherList.tsx
@@ -7,13 +7,13 @@ import {
   getPaginationRowModel,
 } from "@tanstack/react-table";
 
-import { TeacherProfile } from "../../../types/TeacherProfile";
-import { Pagination } from "../../../ui/Pagination";
-import { getTeacherColumns } from "../../../ui/Columns/TeacherColumns";
-import { useEditTeacherModal } from "../../../hooks/custom/useEditTeacherModal";
-import { EditTeacherModal } from "../../../ui/EditTeacherModal";
-import { useGetTeachers } from "../../../hooks/query/useGetTeachers";
-import { usePatchTeacherModal } from "../../../hooks/mutation/usePatchTeacherModal";
+import { TeacherProfile } from "../../../../types/TeacherProfile";
+import { Pagination } from "../../../../ui/Pagination";
+import { getTeacherColumns } from "../../../../ui/Columns/TeacherColumns";
+import { useEditTeacherModal } from "../../../../hooks/custom/useEditTeacherModal";
+import { EditTeacherModal } from "../../../../ui/EditTeacherModal";
+import { useGetTeachers } from "../../../../hooks/query/useGetTeachers";
+import { usePatchTeacherModal } from "../../../../hooks/mutation/usePatchTeacherModal";
 
 function TeacherList() {
   const { data, isLoading, isError } = useGetTeachers();
