+++ app/(route)/(teacher)/layout.tsx
@@ -1,0 +1,7 @@
+export default function TeacherLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <div className="mx-auto max-w-[375px] px-4">{children}</div>;
+}
