+++ app/(route)/(teacher)/teacher/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function TeacherDetail({ params }: { params: { id: string } }) {
+  return <div>teacher</div>;
+}
